+++ NEWS.md
@@ -1,3 +1,9 @@
+# visxhclust 1.0.1
+
+- Remove clusterCrit as a dependency
+- Add clValid as dependency for Dunn index
+- Evaluation module and functions will now only compute silhouette and dunn
+
 # visxhclust 1.0.0
 
 - First version